fix: send device leave event after button released (#566)
Fixes DnD behavior. It always sends a data_device leave event before releasing the dnd grab when the button is released.

+++ src/wayland/data_device/dnd_grab.rs
@@ -188,12 +188,8 @@ impl PointerGrab for DnDGrab {
             if let Some(ref surface) = self.current_focus {
                 if self.data_source.is_some() || self.origin.as_ref().same_client_as(surface.as_ref()) {
                     for device in &seat_data.known_devices {
-                        if device.as_ref().same_client_as(surface.as_ref()) {
-                            if validated {
-                                device.drop();
-                            } else {
-                                device.leave();
-                            }
+                        if device.as_ref().same_client_as(surface.as_ref()) && validated {
+                            device.drop();
                         }
                     }
                 }
@@ -218,6 +214,13 @@ impl PointerGrab for DnDGrab {
             self.icon = None;
             // in all cases abandon the drop
             // no more buttons are pressed, release the grab
+            if let Some(ref surface) = self.current_focus {
+                for device in &seat_data.known_devices {
+                    if device.as_ref().same_client_as(surface.as_ref()) {
+                        device.leave();
+                    }
+                }
+            }
             handle.unset_grab(serial, time);
         }
     }
