+++ src/wayland/data_device/server_dnd_grab.rs
@@ -189,12 +189,8 @@ where
             };
             if let Some(ref surface) = self.current_focus {
                 for device in &seat_data.known_devices {
-                    if device.as_ref().same_client_as(surface.as_ref()) {
-                        if validated {
-                            device.drop();
-                        } else {
-                            device.leave();
-                        }
+                    if device.as_ref().same_client_as(surface.as_ref()) && validated {
+                        device.drop();
                     }
                 }
             }
@@ -213,6 +209,13 @@ where
             }
             // in all cases abandon the drop
             // no more buttons are pressed, release the grab
+            if let Some(ref surface) = self.current_focus {
+                for device in &seat_data.known_devices {
+                    if device.as_ref().same_client_as(surface.as_ref()) {
+                        device.leave();
+                    }
+                }
+            }
             handle.unset_grab(serial, time);
         }
     }
